Remove unused `events` dependency (#51)
`events` is a built-in module[1] in Node.js, so we don't need to
include the `events` library among our dependencies. The Node.js module
resolution algorithm prefers built-in modules over those in
`node_modules`[2], so in practice this library was never used.

It's common for the `events` package to be used in non-Node.js
environments, like the browser. However we aren't producing a build for
non-Node.js environments at the moment, so we have no use for this
package. Any downstream projects can use their own bundlers to
substitute the built-in module with the `events` package.

[1]: https://nodejs.org/docs/latest-v10.x/api/events.html
[2]: https://nodejs.org/docs/latest-v10.x/api/modules.html#modules_all_together