Add `@metamask/auto-changelog`
The `@metamask/auto-changelog` tool has been added to help with
updating and validating the changelog. Changelog validation has been
added to CI, and the pre-existing changelog format has been updated to
comply with the tool's expected changelog format.

The URL format of the `repository` field of `package.json` has been
updated to be compatible with the tool as well.